[codereview] Renamed package and merged two modules
