+++ docker/go-control-plane/controlplane/resource.go
@@ -12,7 +12,7 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-package example
+package controlplane
 
 import (
 	"strconv"
