+++ go.work
@@ -1,0 +1,3 @@
+go 1.20
+
+use ./docker/go-control-plane
